+++ lib/metro/stubs/api_service_stub.dart
@@ -2,14 +2,12 @@ String apiServiceStub({String imports, String apiRequests}) => '''
 import 'dart:convert';
 import 'package:nylo_framework/networking/base_networking.dart';
 import 'package:http/http.dart';
-
 $imports
-
-class BaseApiService extends BaseApi {
+class ApiService extends BaseApi {
   
   var _client;
 
-  BaseApiService() : _client = new Client();
+  ApiService() : _client = new Client();
   
   $apiRequests