+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [0.2.0] - 2020-09-15
+
+* Controllers can now return arguments from routes
+* Dispatch jobs with the new BusQueue
+* Updated router class
+* ApiRender widget for rending Widgets from api futures
+* Removal of unused methods, code clean up, example added
+* New method for storing objects and lists to shared preferences
+* New methods to check app connectivity in networking
+* Metro improvements and bug fixes
+* Bug fixes to project
+
 ## [0.1.0] - 2020-08-23
 
 * Initial release.